+++ src/Models/Hybrid.AuthenticationProfile.psm1
@@ -12,19 +12,27 @@ function New-HybridAuthenticationProfile {
         [string]$DisplayName,
         [string]$DefaultMethod = 'password',
         [string[]]$AuthenticationMethods = @(),
+        [object[]]$AuthenticationMethodDetails = @(),
         [bool]$MfaRegistered = $false,
         [bool]$MfaCapable = $false,
         [bool]$PasswordlessRegistered = $false,
         [bool]$TemporaryAccessPassEligible = $false,
         [string]$AuthenticationStrength = 'Single-factor',
         [string]$ConditionalAccessState = 'Not evaluated',
+        [object[]]$ConditionalAccessDetails = @(),
         [string]$SignInRiskState = 'none',
-        [AllowNull()][datetime]$LastMfaRegistrationDateTime = $null,
-        [AllowNull()][datetime]$LastSuccessfulSignInDateTime = $null,
-        [AllowNull()][datetime]$PasswordLastChangedDateTime = $null,
+        [string]$UserRiskState = '',
+        [string]$RiskLevel = '',
+        [object[]]$RiskDetails = @(),
+        [AllowNull()][Nullable[datetime]]$LastMfaRegistrationDateTime = $null,
+        [AllowNull()][Nullable[datetime]]$LastSuccessfulSignInDateTime = $null,
+        [AllowNull()][Nullable[datetime]]$PasswordLastChangedDateTime = $null,
         [string]$Source = 'MicrosoftGraph',
         [hashtable]$Attributes = @{}
     )
+
+    if ([string]::IsNullOrWhiteSpace($UserRiskState)) { $UserRiskState = $SignInRiskState }
+    if ([string]::IsNullOrWhiteSpace($RiskLevel)) { $RiskLevel = $SignInRiskState }
 
     $profile = [pscustomobject]@{
         PSTypeName = 'Hybrid.AuthenticationProfile'
@@ -32,13 +40,18 @@ function New-HybridAuthenticationProfile {
         DisplayName = $DisplayName
         DefaultMethod = $DefaultMethod
         AuthenticationMethods = @($AuthenticationMethods)
+        AuthenticationMethodDetails = @($AuthenticationMethodDetails)
         MfaRegistered = [bool]$MfaRegistered
         MfaCapable = [bool]$MfaCapable
         PasswordlessRegistered = [bool]$PasswordlessRegistered
         TemporaryAccessPassEligible = [bool]$TemporaryAccessPassEligible
         AuthenticationStrength = $AuthenticationStrength
         ConditionalAccessState = $ConditionalAccessState
+        ConditionalAccessDetails = @($ConditionalAccessDetails)
         SignInRiskState = $SignInRiskState
+        UserRiskState = $UserRiskState
+        RiskLevel = $RiskLevel
+        RiskDetails = @($RiskDetails)
         LastMfaRegistrationDateTime = $LastMfaRegistrationDateTime
         LastSuccessfulSignInDateTime = $LastSuccessfulSignInDateTime
         PasswordLastChangedDateTime = $PasswordLastChangedDateTime